refactor: rename ibuffer.h -> ibuffer_avr.h
