+++ compiler/main/x86_code_gen/cg_x86_single_function.c
@@ -15,7 +15,7 @@
 #include "ast/ast.h"
 #include "ast/ast_declare.h"
 
-#include "ibuffer/ibuffer.h"
+#include "ibuffer/ibuffer_avr.h"
 
 #include "tables/symtable/symtable.h"
 #include "cg_x86_single_function.h"
