+++ ibuffer/ibuffer_write.h
@@ -1,8 +1,9 @@
 #pragma once
 
+#include <stdio.h>
+
 #include <inttypes.h>
 
-#include "ibuffer.h"
 #include "ikey.h"
 
 void ibu_write_instr(enum IKEY key, int32_t x1, int32_t x2, int32_t x3, char* str, char* comment, FILE* fout);
